added basic home for each role, refactor of the user for returning response for each DAO action adding tests in progress..

+++ com.team6.project/src/main/java/com/team6/project/dao/jpa/JPAUserDAO.java
@@ -9,9 +9,10 @@ import javax.persistence.PersistenceContext;
 import com.team6.project.dao.UserDAO;
 import com.team6.project.entities.Response;
 import com.team6.project.entities.User;
+
 @Stateless
 @Local
-@DeclareRoles({"administrator"})
+@DeclareRoles({ "administrator" })
 public class JPAUserDAO implements UserDAO {
 
     @PersistenceContext
@@ -20,11 +21,10 @@ public class JPAUserDAO implements UserDAO {
     @Override
     public Response addUser(User user) {
         Response response = new Response();
-        if (em.find(User.class, user.getUserId()) == null) {
+        if (em.find(User.class, user.getKey()) == null) {
             em.persist(user);
             response.setStatus(Response.Status.OK);
-        }
-        else{
+        } else {
             response.setStatus(Response.Status.ERROR);
             response.setDescription("User already exists");
         }
@@ -32,19 +32,35 @@ public class JPAUserDAO implements UserDAO {
     }
 
     @Override
-    public void updateUser(User user) {
-        em.merge(user);
+    public Response updateUser(User user) {
+        Response response = new Response();
+        if (em.find(User.class, user.getKey()) != null) {
+            em.merge(user);
+            response.setStatus(Response.Status.OK);
+        } else {
+            response.setStatus(Response.Status.NOT_FOUND);
+            response.setDescription("User not found");
+        }
+        return response;
     }
 
     @Override
     public User getUserByKey(String userId) {
-       return em.find(User.class, userId);
+        return em.find(User.class, userId);
 
     }
 
     @Override
-    public void daleteUser(User user) {
-        em.remove(user);
+    public Response deleteUser(User user) {
+        Response response = new Response();
+        if (em.find(User.class, user.getKey()) != null) {
+            em.remove(user);
+            response.setStatus(Response.Status.OK);
+        } else {
+            response.setStatus(Response.Status.NOT_FOUND);
+            response.setDescription("User not found");
+        }
+        return response;
 
     }
